pyproject.toml
[project]
name = "scrapddl"
version = "0.1.0"
description = ""
authors = [
    {name = "scrapddl", email = "scrapddl.q4g1w@simplelogin.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "flask>=2.2.2",
    "lxml>=6.0.2",
    "pyopenssl>=23.0.0",
    "requests>=2.28.2",
    "scrap-imdb>=0.1.8",
    "python-slugify>=8.0.0",
    "cachelib>=0.10.2",
    "cloudscraper>=1.2.68",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.2.1",
    "ipdb>=0.13.11",
    "mypy>=1.8.0",
    "types-requests>=2.31.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.mypy]
python_version = "3.11"
warn_return_any = false
warn_unused_ignores = false
disallow_untyped_defs = true
disallow_incomplete_defs = true
ignore_missing_imports = true
exclude = [
    "scrapddl/tests/",
]
